+++ ui-components/src/helpers/filters.ts
@@ -13,6 +13,8 @@ function getComparisonForFilter(filterKey: ListingFilterKeys) {
       return EnumListingFilterParamsComparison["="]
     case ListingFilterKeys.bedrooms:
       return EnumListingFilterParamsComparison[">="]
+    case ListingFilterKeys.zipcode:
+      return EnumListingFilterParamsComparison["IN"]
     default: {
       const _exhaustiveCheck: never = filterKey
       return _exhaustiveCheck